+++ tests/recipes/recipes.css.ts
@@ -32,3 +32,23 @@ export const basic = recipe({
     },
   ],
 });
+
+export const empty = recipe({});
+
+export const definedStringBase = recipe({
+  base: 'definedStringBase',
+  variants: {
+    variant: {
+      simple: 'simple-one',
+    },
+  },
+});
+
+export const definedStringBaseArray = recipe({
+  base: ['definedStringBaseInArray_1', 'definedStringBaseInArray_2'],
+  variants: {
+    variant: {
+      simple: ['simple-one', 'simple-two'],
+    },
+  },
+});